Confirmation prompt & its anim

+++ Assets/Scripts/Menus/ButtonsScript.cs
@@ -6,10 +6,20 @@ using UnityEngine.SceneManagement;
 public class ButtonsScript : MonoBehaviour
 {
     private string root;
-
+    private bool _HasSave;
     private void Start()
     {
         root = Application.persistentDataPath;
+        var files = from file in Directory.EnumerateFiles(root) select file;
+        foreach (var file in files)
+        {
+            if (file.EndsWith(".json"))
+            {
+                _HasSave = true;
+                return;
+            }
+        }
+        _HasSave = false;
     }
     public void Quit()
     {
@@ -18,29 +28,43 @@ public class ButtonsScript : MonoBehaviour
 
     public void Load()
     {
-        var files = from file in Directory.EnumerateFiles(root) select file;
-        foreach(var file in files)
+        if (_HasSave)
         {
-            if (file.EndsWith(".json"))
-            {
-                SceneManager.LoadSceneAsync(1);
-                return;
-            }
+            SceneManager.LoadSceneAsync(1);
+            return;
         }
     }
     public void Begin()
     {
-        var files = from file in Directory.EnumerateFiles(root) select file;
-        if (files.Count() > 0)
+        //shows confirmation prompt in case the user already has a save file
+        if (_HasSave)
         {
-            foreach (var file in files)
+            ConfirmationPrompt.Instance.OpenConfirmationPrompt(
+            () =>
             {
-                if(!file.EndsWith(".log"))
-                    File.Delete(file);
-            }
+                var files = from file in Directory.EnumerateFiles(root) select file;
+                if (files.Count() > 0)
+                {
+                    foreach (var file in files)
+                    {
+                        if (!file.EndsWith(".log"))
+                            File.Delete(file);
+                    }
+                }
+                if (Directory.Exists("Game")) Directory.Delete("Game", true);
+                SceneManager.LoadSceneAsync(1);
+                ConfirmationPrompt.Instance.ChangePromptState();
+            }, 
+            () => 
+            {
+                ConfirmationPrompt.Instance.ChangePromptState();
+            }, 
+            "Si vous continuez, toute progression sera effacée !");
         }
-        if (Directory.Exists("Game")) Directory.Delete("Game",true);
-        SceneManager.LoadSceneAsync(1);
+        else
+        {
+            SceneManager.LoadSceneAsync(1);
+        }
     }
     public void Settings()
     {
